+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 
-> 1.1.0 (Intial)
+> 1.1.1
+- Fix multi cursors in same line just insert semicolon without smart placement
+- Add go as excluded language
+- Update Gifs
 
+> 1.1.0 (Intial)
 - Added Multi-line support
 - Added Disabled languages
 - Added fallback on cstyle insertion